Enabled server port configuration.
This commit modifies the server (src/server_main.cc) such that it configures the port it listens to from the configure file passed in, instead of having to manually pass in the port number as a parameter each time starting up the server.

The CMakelist.txt has been updated to include a config parser (src/utils/config_parser.cc), which parses the config file under the Nginx format. A GetPort function has been added to the config parser to allow the server to extract port information from the config file passed in.

This is a preparation for transferring the server setup to utilization of config files, which would be more convenient than passing in parameters to the server program. Since in the future there would be more parameters to the server in starting phase.

Tests have been added under tests/config_parser_test.cc to test the functionality of port parsing from config files.

Change-Id: I445da7ee98b6090e400906dd595248a1b8eff741