Refactor code for reading request body
When writing unit test for read_request_body
function, it turns out the way it is implemented
is hard to test because it depends on socket,
and mocking a socket is a pain.

Change the function signature such that it is socket
independent. Will submit another review for unit test
based on this change.

Change-Id: I6adb4059a62bc52ec38365a8e28cd33f1b6e4286