Add simple integration test
Add some skeleton code for integration test.
This test checks if the server responds to client
as expected. In particular, I added a test to
check if the response to GET request is correct.

To just run the integration test, just type:
	$ ./integration_test.sh

in the build folder after make.

We still need to add more test cases for integration test,
but this will get us started.

Change-Id: I294e3f9827d1480e10180f46790e11cf7dea28c7