Fix server test from failing on Google Cloud
One of the server test was failing because the
client tries to connect to the server before
it starts up. This was not an issue on local
development environment because server starts up
immediately even though it was in another thread.

However, in the production environment, either
because there is a lof of resources competing for
CPU or lack of CPU (we are using the cheapest machine),
the server thread is not executed immediately, and thus
results in a situation where the client tries to connect
to a server that does not exist.

Make current thread sleep for one second after creating
the server thread solves this problem.

Change-Id: I02897bcff0c2e5e5a1193f77003bacf2be9c4cf6