This change is mainly to add tests for most the parts we have modified this time.
The parts of tests for static requests, which is new this assignment, include echo requests, mimetypes, handler factory, session manager, and integration tests.

The session manager test is not working. The handler factory test is working. However, if you uncomment the test for handler factory test to test the static req handler in the handler factory, it would work if you use "bin/request_handler_facotry_test" individually, but if use "make test", it won't work. With the current version, the handler facotry test is working fine.

Change-Id: I4cbd4f7147b6911b65c30d8a4d1818b1ca42f8e6