Reflect the changes of Common API
Details of the change can be found here:
https://www.cs130.org/assignments/6/#updates-from-class-discussion-2020-05-11

All the methods and fields that voiolate the common API
are removed.

Patchset 2:
Due to change of API, the number of files in http/server directory
was too many and it was hard to manage. Add more directories for
better organization.

http --- server --- server_main.cc
      |          |- server.cc
      |
      |- session --- session.cc
      |           |- session_manager.cc
      |
      |- request --- request.cc
      |           |- request_parser.cc
      |
      |- request_handler --- request_handler.cc etc
      |
      |- header.cc
      |- method_type.cc
      |- ...

Namespace needs to be updated but that should be a different
review since this review is already big enough.


Change-Id: Iae8eda170057a121c7a0c0692cdc8d96b1038bce